pyproject.toml
[project]
name = "zndraw-utils"
version = "0.1.0"
description = "ZnDraw Extensions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ase>=3.23.0",
    "mace-torch>=0.3.9",
    "rdkit2ase>=0.1.4",
    "tqdm>=4.67.1",
    "typer>=0.13.1",
    "zndraw>=0.5.7",
]

[dependency-groups]
dev = [
    "mace-torch>=0.3.9",
    "ruff>=0.8.3",
]

[project.scripts]
zndraw-register = "zndraw_utils.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
